requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask[async]==3.0.0
Flask-Cors==4.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
zipp==3.17.0
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
flasgger==0.9.7.1
python-socketio==5.5.0
websockets==10.1
redis==4.0.2
aioredis==2.0.0